requirements.txt
aiogram>=3.0,<4
httpx>=0.27
openai>=1.0
python-dotenv>=1.0
ruff>=0.4

# Backend (FastAPI, async SQLAlchemy, Alembic; см. docs/adr/adr-002-backend-http-orm.md)
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pytest>=8.0
pytest-asyncio>=0.24
psycopg[binary]>=3.2
